Correcciones de Variables y relación entre laboratorio y examen

+++ src/controlador/LaboratorioJpaController.java
@@ -5,6 +5,7 @@
  */
 package controlador;
 
+import controlador.exceptions.IllegalOrphanException;
 import controlador.exceptions.NonexistentEntityException;
 import java.io.Serializable;
 import javax.persistence.Query;
@@ -28,9 +29,10 @@ public class LaboratorioJpaController implements Serializable {
     public LaboratorioJpaController(EntityManagerFactory emf) {
         this.emf = emf;
     }
-    private EntityManagerFactory emf = Persistence.createEntityManagerFactory("SistemaHospitalarioPU");
+    private EntityManagerFactory emf = null;
 
     public LaboratorioJpaController() {
+         emf = Persistence.createEntityManagerFactory("SistemaHospitalarioPU");
     }
 
     public EntityManager getEntityManager() {
@@ -38,23 +40,24 @@ public class LaboratorioJpaController implements Serializable {
     }
 
     public void create(Laboratorio laboratorio) {
-        if (laboratorio.getListaExamen() == null) {
-            laboratorio.setListaExamen(new ArrayList<Examen>());
-        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            List<Examen> attachedListaExamen = new ArrayList<Examen>();
-            for (Examen listaExamenExamenToAttach : laboratorio.getListaExamen()) {
-                listaExamenExamenToAttach = em.getReference(listaExamenExamenToAttach.getClass(), listaExamenExamenToAttach.getId_examen());
-                attachedListaExamen.add(listaExamenExamenToAttach);
+            Examen examen = laboratorio.getExamen();
+            if (examen != null) {
+                examen = em.getReference(examen.getClass(), examen.getId_examen());
+                laboratorio.setExamen(examen);
             }
-            laboratorio.setListaExamen(attachedListaExamen);
             em.persist(laboratorio);
-            for (Examen listaExamenExamen : laboratorio.getListaExamen()) {
-                listaExamenExamen.getListaLab().add(laboratorio);
-                listaExamenExamen = em.merge(listaExamenExamen);
+            if (examen != null) {
+                Laboratorio oldLaboratorioOfExamen = examen.getLaboratorio();
+                if (oldLaboratorioOfExamen != null) {
+                    oldLaboratorioOfExamen.setExamen(null);
+                    oldLaboratorioOfExamen = em.merge(oldLaboratorioOfExamen);
+                }
+                examen.setLaboratorio(laboratorio);
+                examen = em.merge(examen);
             }
             em.getTransaction().commit();
         } finally {
@@ -64,33 +67,37 @@ public class LaboratorioJpaController implements Serializable {
         }
     }
 
-    public void edit(Laboratorio laboratorio) throws NonexistentEntityException, Exception {
+    public void edit(Laboratorio laboratorio) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
             Laboratorio persistentLaboratorio = em.find(Laboratorio.class, laboratorio.getId_laboratorio());
-            List<Examen> listaExamenOld = persistentLaboratorio.getListaExamen();
-            List<Examen> listaExamenNew = laboratorio.getListaExamen();
-            List<Examen> attachedListaExamenNew = new ArrayList<Examen>();
-            for (Examen listaExamenNewExamenToAttach : listaExamenNew) {
-                listaExamenNewExamenToAttach = em.getReference(listaExamenNewExamenToAttach.getClass(), listaExamenNewExamenToAttach.getId_examen());
-                attachedListaExamenNew.add(listaExamenNewExamenToAttach);
+            Examen examenOld = persistentLaboratorio.getExamen();
+            Examen examenNew = laboratorio.getExamen();
+            List<String> illegalOrphanMessages = null;
+            if (examenOld != null && !examenOld.equals(examenNew)) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("You must retain Examen " + examenOld + " since its laboratorio field is not nullable.");
             }
-            listaExamenNew = attachedListaExamenNew;
-            laboratorio.setListaExamen(listaExamenNew);
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            if (examenNew != null) {
+                examenNew = em.getReference(examenNew.getClass(), examenNew.getId_examen());
+                laboratorio.setExamen(examenNew);
+            }
             laboratorio = em.merge(laboratorio);
-            for (Examen listaExamenOldExamen : listaExamenOld) {
-                if (!listaExamenNew.contains(listaExamenOldExamen)) {
-                    listaExamenOldExamen.getListaLab().remove(laboratorio);
-                    listaExamenOldExamen = em.merge(listaExamenOldExamen);
+            if (examenNew != null && !examenNew.equals(examenOld)) {
+                Laboratorio oldLaboratorioOfExamen = examenNew.getLaboratorio();
+                if (oldLaboratorioOfExamen != null) {
+                    oldLaboratorioOfExamen.setExamen(null);
+                    oldLaboratorioOfExamen = em.merge(oldLaboratorioOfExamen);
                 }
-            }
-            for (Examen listaExamenNewExamen : listaExamenNew) {
-                if (!listaExamenOld.contains(listaExamenNewExamen)) {
-                    listaExamenNewExamen.getListaLab().add(laboratorio);
-                    listaExamenNewExamen = em.merge(listaExamenNewExamen);
-                }
+                examenNew.setLaboratorio(laboratorio);
+                examenNew = em.merge(examenNew);
             }
             em.getTransaction().commit();
         } catch (Exception ex) {
@@ -109,7 +116,7 @@ public class LaboratorioJpaController implements Serializable {
         }
     }
 
-    public void destroy(Long id) throws NonexistentEntityException {
+    public void destroy(Long id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -121,10 +128,16 @@ public class LaboratorioJpaController implements Serializable {
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The laboratorio with id " + id + " no longer exists.", enfe);
             }
-            List<Examen> listaExamen = laboratorio.getListaExamen();
-            for (Examen listaExamenExamen : listaExamen) {
-                listaExamenExamen.getListaLab().remove(laboratorio);
-                listaExamenExamen = em.merge(listaExamenExamen);
+            List<String> illegalOrphanMessages = null;
+            Examen examenOrphanCheck = laboratorio.getExamen();
+            if (examenOrphanCheck != null) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Laboratorio (" + laboratorio + ") cannot be destroyed since the Examen " + examenOrphanCheck + " in its examen field has a non-nullable laboratorio field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(laboratorio);
             em.getTransaction().commit();
